DLF GH63 Sector 63 AT Gurgaon
DLF GH63 Sector 63 AT Gurgaon | DLF Crest Level specifications

DLF GH 63 Gurgaon is a residential project that comes with lavish living experiences. These are ultra-opulent apartments that will change your standard of living in a grand way. With lavishness in every nook and corner of the project by renowned designers and architects, it stands out as a pure luxury.

The project promises to give you a lifetime of happiness and a resplendent lifestyle that matches your style and personality. A true reflection of your style in its purest form with luxury and class, even in the smallest details. DLF GH63 Apartments in Gurugram is absolutely unmatched and unique which resonates with the Luxury living life.

It stands tall in its name where life begins towards a better and brighter future. Total 1250 Apartments (6 Towers) Stilt + 35 floors are wonderfully stylish, with heights matching the globally acclaimed designs. These are the most luxurious residential apartments that offer vibrant living in the entire area. The height of thoughts as well as the excellent state of mind, body, and spirit make you truly gorgeous. These well designed All 4BHK + Utility (3400 Sqft) fully loaded apartments. DLF GH63 Sector 63 Floor plans have impressive designs to accommodate every need and home decor. It seeks maximum space and comfort for all residents giving a sense of freedom.

Key Highlights of this project: 

➡️ DLF Crest Level specifications.
➡️ Super Luxury lobbies like DLF The Camellias.
➡️ 80% green and open Area. 
➡️ 1.5 Lac Square feet Luxury Club House.
➡️ Tentative pricing:- ₹20000-22000/- per square feet. 
➡️ Possession in 2026.

It takes you to the seventh heaven with a vast presence of luxury and huge green. See the beautiful scenery every morning and stare at the stars at night to make life more worthwhile. Landscaped podium gardens free from vehicles, an air-conditioned gymnasium, a multipurpose hall, an automatic lift and much more add a lot to the monotonous life. In a community with secure gates, feel the sense of security round the clock and worry-free for your kids and even elders as they are going to make new friends.

DLF GH 63 Gurgaon Apartments is an ideal place to stay close to an affluent neighborhood and well-established social infrastructure. It offers seamless connectivity to major destinations across the city which makes it more attractive for home buyers.

For More Information
Visit Us: - https://bit.ly/3fWKF3Z